Started MCTS implementation and renamed some files/classes
